+++ neo/d3xp/physics/Physics_Player.h
@@ -113,7 +113,7 @@ public:
 
 public:	// common physics interface
 	bool					Evaluate( ID_TIME_T timeStepMSec, ID_TIME_T endTimeMSec ) override;
-	bool					Interpolate( const float fraction ) override;
+	bool					Interpolate( const double fraction ) override;
 	void					UpdateTime( ID_TIME_T endTimeMSec ) override;
 	ID_TIME_T				GetTime() const override;
