+++ neo/framework/Console.h
@@ -81,7 +81,7 @@ public:
 
 	virtual void	PrintOverlay( idOverlayHandle & handle, justify_t justify, VERIFY_FORMAT_STRING const char * text, ... ) = 0;
 
-	virtual idDebugGraph *	CreateGraph( int numItems ) = 0;
+	virtual idDebugGraph *	CreateGraph( size_t numItems ) = 0;
 	virtual void			DestroyGraph( idDebugGraph * graph ) = 0;
 };
