+++ neo/renderer/Image.h
@@ -316,7 +316,7 @@ public:
 
 extern idImageManager	*globalImages;		// pointer to global list for the rest of the system
 
-int MakePowerOfTwo( int num );
+int MakePowerOfTwo( size_t num );
 
 /*
 ====================================================================
